+++ reps/api/__init__.py
@@ -6,7 +6,7 @@ modules; this package is a thin facade.
 
 from reps.api.example import Example, Prediction
 from reps.api.model import Model, ModelKwargs
-from reps.api.objective import LLMJudge, Objective
+from reps.api.objective import LLMJudge, Objective, PromptObjective
 from reps.api.optimizer import Optimizer
 from reps.api.result import OptimizationResult
 from reps.evaluation_result import EvaluationResult
@@ -16,6 +16,7 @@ __all__ = [
     "Prediction",
     "Objective",
     "LLMJudge",
+    "PromptObjective",
     "Model",
     "ModelKwargs",
     "Optimizer",